Create Responsive Design
Responsive Design helps resize images, fits them to the screen, and helps adjust to High Density Dot Per Inch Display (HD-DPI Displays) per device. Each device has different amounts of pixels that it is able to display, therefore depending on which device you are using to view the image, the amount of pixels used for the image can be changed. This code is noted through.

        function resizeRendererToDisplaySize(renderer) {
            const canvas = renderer.domElement;
            const pixelRatio = window.devicePixelRatio;
            const width = canvas.clientWidth * pixelRatio | 0;
            const height = canvas.clientHeight * pixelRatio | 0;
            const needResize = canvas.width !== width || canvas.height !== height;
            if (needResize) {
                renderer.setSize(width, height, false)
            }
            return needResize;
        }